Added option for enhanced accuracy for scenery and other developers
Enhanced accuracy shows more decimal places for measurement lines, flight plan lines and range
rings, if enabled.
Using double precision for range rings now
Optimization in Unit: more inlines.
Fixed issues with inaccurate display of range ring distances.
Now using correct locale dependent sign
#1249

+++ src/common/mapmarkertypes.h
@@ -151,7 +151,7 @@ struct RangeMarker
   QString displayText() const;
 
   QString text; /* Text to display like VOR name and frequency */
-  QList<float> ranges; /* Range ring list (NM) */
+  QList<double> ranges; /* Range ring list (NM) */
   MapType navType; /* VOR, NDB, AIRPORT, etc. */
   QColor color;
 };
